+++ src/main/java/pl/nankiewic/fleetappbackend/Service/RefuelingService.java
@@ -4,10 +4,10 @@ import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import pl.nankiewic.fleetappbackend.DTO.RefuelingDTO;
 import pl.nankiewic.fleetappbackend.Entity.VehicleRefueling;
+import pl.nankiewic.fleetappbackend.Exception.PermissionDeniedException;
 import pl.nankiewic.fleetappbackend.Mapper.RefuelingMapper;
 import pl.nankiewic.fleetappbackend.Repository.VehicleRefuelingRepository;
 import pl.nankiewic.fleetappbackend.Repository.UserRepository;
-import pl.nankiewic.fleetappbackend.Repository.VehicleRepository;
 
 import javax.persistence.EntityNotFoundException;
 
@@ -15,34 +15,38 @@ import javax.persistence.EntityNotFoundException;
 @Service
 public class RefuelingService {
 
-    private final VehicleRepository vehicleRepository;
+    private final CheckExistAndPermissionComponent checkExistAndPermissionComponent;
     private final VehicleRefuelingRepository refuelingRepository;
     private final UserRepository userRepository;
     private final RefuelingMapper refuelingMapper;
 
     public void createVehicleRefueling(RefuelingDTO refuelingDTO, String email) {
-        VehicleRefueling vehicleRefueling = refuelingMapper.refuelingDtoToVehicleRefuelingEntity(refuelingDTO);
-        vehicleRefueling.setUser(userRepository.findUserByEmail(email));
-        refuelingRepository.save(vehicleRefueling);
+        if (checkExistAndPermissionComponent.accessToVehicle(email, refuelingDTO.getVehicleId())){
+            VehicleRefueling vehicleRefueling = refuelingMapper.refuelingDtoToVehicleRefuelingEntity(refuelingDTO);
+            vehicleRefueling.setUser(userRepository.findUserByEmail(email));
+            refuelingRepository.save(vehicleRefueling);
+        }  else throw new PermissionDeniedException();
     }
 
-    public void updateVehicleRefueling(RefuelingDTO refuelingDTO) {
-        VehicleRefueling vehicleRefueling = refuelingRepository.findById(refuelingDTO.getId()).orElseThrow(
-                () -> new EntityNotFoundException("Refueling not found"));
-        refuelingMapper.updateVehicleRepairFromDto(vehicleRefueling, refuelingDTO);
-        refuelingRepository.save(vehicleRefueling);
+    public void updateVehicleRefueling(RefuelingDTO refuelingDTO, String email) {
+        if (checkExistAndPermissionComponent.accessToRefueling(email, refuelingDTO.getVehicleId())) {
+            VehicleRefueling vehicleRefueling = refuelingRepository.findById(refuelingDTO.getId()).orElseThrow(
+                    () -> new EntityNotFoundException("Refueling not found"));
+            refuelingMapper.updateVehicleRepairFromDto(vehicleRefueling, refuelingDTO);
+            refuelingRepository.save(vehicleRefueling);
+        } else throw new PermissionDeniedException();
     }
 
-    public RefuelingDTO getRefuelingById(Long id) {
-        if (refuelingRepository.existsById(id)) {
+    public RefuelingDTO getRefuelingById(Long id, String email) {
+        if (checkExistAndPermissionComponent.accessToRefueling(email, id)) {
             return refuelingRepository.findVehicleRefuelingById(id);
-        } else throw new EntityNotFoundException("Refueling not found");
+        } else throw new PermissionDeniedException();
     }
 
-    public Iterable<RefuelingDTO> getRefuelingByVehicle(Long id) {
-        if (vehicleRepository.existsById(id)) {
+    public Iterable<RefuelingDTO> getRefuelingByVehicle(Long id, String email) {
+        if (checkExistAndPermissionComponent.accessToVehicle(email, id)) {
             return refuelingRepository.findRefuelingListByVehicle(id);
-        } else throw new EntityNotFoundException("Vehicle not found");
+        } else throw new PermissionDeniedException();
     }
 
     public Iterable<RefuelingDTO> getRefuelingByUser(String email) {
@@ -57,14 +61,18 @@ public class RefuelingService {
         } else throw new EntityNotFoundException("User not found");
     }
 
-    public Iterable<RefuelingDTO> getRefuelingByUserAndVehicle(Long userId, Long vehicleId) {
-        if (userRepository.existsById(userId) && vehicleRepository.existsById(vehicleId)) {
-            return refuelingRepository.findAllByVehicleAndUser(vehicleId, userId);
-        } else throw new EntityNotFoundException("User or Vehicle not found");
+    public Iterable<RefuelingDTO> getRefuelingByUserAndVehicle(Long userId, Long vehicleId, String email) {
+        if (checkExistAndPermissionComponent.accessToVehicle(email, vehicleId)) {
+            if (userRepository.existsById(userId)) {
+                return refuelingRepository.findAllByVehicleAndUser(vehicleId, userId);
+            } else throw new EntityNotFoundException("User not found");
+        } else throw new PermissionDeniedException();
     }
 
-    public void deleteRefuelingById(Long id) {
-        refuelingRepository.deleteById(id);
+    public void deleteRefuelingById(Long id, String email) {
+        if (checkExistAndPermissionComponent.accessToRefueling(email, id)) {
+            refuelingRepository.deleteById(id);
+        } else throw new PermissionDeniedException();
     }
 
 }
