+++ src/main/java/pl/nankiewic/fleetappbackend/Service/UseService.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Service;
 import pl.nankiewic.fleetappbackend.DTO.UseDTO;
 import pl.nankiewic.fleetappbackend.Entity.Vehicle;
 import pl.nankiewic.fleetappbackend.Entity.VehicleUse;
+import pl.nankiewic.fleetappbackend.Exception.PermissionDeniedException;
 import pl.nankiewic.fleetappbackend.Mapper.UseMapper;
 import pl.nankiewic.fleetappbackend.Repository.UserRepository;
 import pl.nankiewic.fleetappbackend.Repository.VehicleRepository;
@@ -16,59 +17,65 @@ import javax.persistence.EntityNotFoundException;
 @Service
 public class UseService {
 
+    private final CheckExistAndPermissionComponent checkExistAndPermissionComponent;
     private final VehicleUseRepository vehicleUseRepository;
     private final VehicleRepository vehicleRepository;
     private final UserRepository userRepository;
     private final UseMapper useMapper;
 
     public void createVehicleUse(UseDTO use, String email) {
-
-        VehicleUse vehicleUse = useMapper.vehicleUseDtoToEntity(use);
-        addMileageToVehicle(use.getVehicleId(), use.getTrip());
-        vehicleUse.setUser(userRepository.findUserByEmail(email));
-        vehicleUseRepository.save(vehicleUse);
+        if (checkExistAndPermissionComponent.accessToVehicle(email, use.getVehicleId())) {
+            VehicleUse vehicleUse = useMapper.vehicleUseDtoToEntity(use);
+            addMileageToVehicle(use.getVehicleId(), use.getTrip());
+            vehicleUse.setUser(userRepository.findUserByEmail(email));
+            vehicleUseRepository.save(vehicleUse);
+        } else throw new PermissionDeniedException();
     }
 
-    public void updateVehicleUse(UseDTO useDTO) {
-
-        VehicleUse vehicleUse = vehicleUseRepository.findById(useDTO.getId()).orElseThrow(
-                () -> new EntityNotFoundException("Bład przetwarzania"));
-        Short currentTripMileageFromDatabase = vehicleUse.getTrip();
-        Short newValueOfMilTrip = useDTO.getTrip();
-        useMapper.updateVehicleUseFromDto(vehicleUse, useDTO);
-        recalculateMileageToVehicle(useDTO.getVehicleId(), currentTripMileageFromDatabase, newValueOfMilTrip);
-        vehicleUseRepository.save(vehicleUse);
+    public void updateVehicleUse(UseDTO useDTO, String email) {
+        if (checkExistAndPermissionComponent.accessToUse(email, useDTO.getId())) {
+            VehicleUse vehicleUse = vehicleUseRepository.findById(useDTO.getId()).orElseThrow(
+                    () -> new EntityNotFoundException("Use not found"));
+            Short currentTripMileageFromDatabase = vehicleUse.getTrip();
+            Short newValueOfMilTrip = useDTO.getTrip();
+            useMapper.updateVehicleUseFromDto(vehicleUse, useDTO);
+            recalculateMileageToVehicle(useDTO.getVehicleId(), currentTripMileageFromDatabase, newValueOfMilTrip);
+            vehicleUseRepository.save(vehicleUse);
+        } else throw new PermissionDeniedException();
     }
 
-    public Iterable<UseDTO> getUseByVehicle(Long id) {
-
-        if (vehicleRepository.existsById(id)) {
+    public Iterable<UseDTO> getUseByVehicle(Long id, String email) {
+        if (checkExistAndPermissionComponent.accessToVehicle(email, id)) {
             return vehicleUseRepository.findAllByVehicleId(id);
-        } else throw new EntityNotFoundException();
+        } else throw new PermissionDeniedException();
     }
 
-    public UseDTO getUseByUseId(Long id) {
-        return vehicleUseRepository.findByUseId(id);
+    public UseDTO getUseByUseId(Long id, String email) {
+        if (checkExistAndPermissionComponent.accessToUse(email, id)) {
+            return vehicleUseRepository.findByUseId(id);
+        } else throw new PermissionDeniedException();
     }
 
     public Iterable<UseDTO> getUseByUser(String email) {
         return vehicleUseRepository.findAllByUserId(userRepository.findUserByEmail(email).getId());
     }
 
-    public Iterable<UseDTO> getUseByUserAndVehicle(Long userId, Long vehicleId) {
-        if (userRepository.existsById(userId) && vehicleRepository.existsById(vehicleId)) {
+    public Iterable<UseDTO> getUseByUserAndVehicle(Long userId, Long vehicleId, String email) {
+        if (checkExistAndPermissionComponent.accessToVehicle(email, vehicleId)) {
             return vehicleUseRepository.findAllByVehicleAndUser(vehicleId, userId);
-        } else throw new EntityNotFoundException("Nie znaleziono zasobu pojazd lub użytkownik");
+        } else throw new PermissionDeniedException();
     }
 
-    public void deleteUseById(Long id) {
-        VehicleUse use = vehicleUseRepository.findById(id).orElseThrow(
-                () -> new EntityNotFoundException("Bład przetwarzania"));
-        Vehicle vehicle = use.getVehicle();
-        int mil = Integer.parseInt(vehicle.getMileage()) - use.getTrip();
-        vehicle.setMileage(Integer.toString(mil));
-        vehicleRepository.save(vehicle);
-        vehicleUseRepository.deleteById(id);
+    public void deleteUseById(Long id, String email) {
+        if (checkExistAndPermissionComponent.accessToUse(email, id)) {
+            VehicleUse use = vehicleUseRepository.findById(id).orElseThrow(
+                    () -> new EntityNotFoundException("Bład przetwarzania"));
+            Vehicle vehicle = use.getVehicle();
+            int mil = Integer.parseInt(vehicle.getMileage()) - use.getTrip();
+            vehicle.setMileage(Integer.toString(mil));
+            vehicleRepository.save(vehicle);
+            vehicleUseRepository.deleteById(id);
+        } else throw new PermissionDeniedException();
     }
 
     private void addMileageToVehicle(Long vehicleId, Short trip) {
